refactor: renamed language files to be more broad
